Finished initial subdivision code and added two examples

+++ CogoTestBed/CogoTestBed/Node.cs
@@ -10,9 +10,9 @@ namespace CogoTestBed
     {
         #region Properties
 
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         #endregion
